Added the core of fremit
This is a basic start, but represents the type of API I want, and
the theme of a singleton event stream for the app. I'd like to add
more functions to Stream, mostly, but this is what I've needed to
get up and running.

Also, testing is dismal right now, but at least there's a start.